chatbot/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
pydantic==2.11.7
google-generativeai==0.8.5
youtube-transcript-api==0.6.2
python-dotenv==1.1.1
firebase-admin==6.5.0